Fix pointer value of object to string.

+++ src/bndbox.cpp
@@ -86,7 +86,7 @@ mrb_value siren_bndbox_to_s(mrb_state* mrb, mrb_value self)
   b->IsOpenZmax() ? snprintf(szmax, s, "%s", "inf") : snprintf(szmax, s, "%f", zmax);
   snprintf(str, sizeof(str),
       "#<BndBox:0x%x xmin=%s, ymin=%s, zmin=%s, xmax=%s, ymax=%s, zmax=%s>",
-      (int)mrb_cptr(self), sxmin, symin, szmin, sxmax, symax, szmax);
+      (uintptr_t)mrb_cptr(self), sxmin, symin, szmin, sxmax, symax, szmax);
   return mrb_str_new_cstr(mrb, str);
 }
